Added dotdeb repository so that Debian can find php5-fpm.
Ensured that the /run/shm directory is created otherwise the fpm
process fails to start.

Added the Augeua tool set so that the default timezone can be set,
otherwise Symfony complains very loudly about it. Used these tools to
update the default.timezome config option for both the CLI & fpm
versions of PHP.