Fixed provisioning problems for Debian boxes.
Mainly due to Puppet trying to install from the Debian repo before it's
been added & refreshed. I've added a requirement on the "after_apt"
anchor for the bits that were failing, this will still get executed on
non-debian hosts, but needs testing.

I also changed php-apc to php5-apc, if this module name is specific to
Debian then we can add something to the $osfamily switch to make it
work, again a test on a non-Debian VM will show us if we need that or
not.